LayoutTests/crypto/subtle/aes-cbc-import-key-unwrap-spki-key-expected.txt

Test unwrapping a SPKI key with AES-CBC using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new Uint8Array(unwrappedKey) is spkiKey
PASS successfullyParsed is true

TEST COMPLETE
